Brownie/requirements.txt
flask
eth-brownie
gradio
numpy
matplotlib
